+++ packages/shared/lib/index.ts
@@ -26,6 +26,7 @@ export * from './services/sync/job.service.js';
 export * from './services/sync/config/config.service.js';
 export * from './services/sync/config/endpoint.service.js';
 export * from './services/sync/config/deploy.service.js';
+export * from './services/sync/utils.js';
 export * from './services/endUser.service.js';
 export * from './services/onboarding.service.js';
 export * from './services/invitations.js';
